+++ CHANGES.rst
@@ -4,6 +4,8 @@
 - added :class:`udev.Monitor` to support event monitoring
 - added :meth:`udev.Device.asbool` and :meth:`udev.Device.asint`
 - removed type magic in :meth:`udev.Device.__getitem__`
+- added :mod:`qudev` to for PyQt4 integration
+
 
 0.1 (May 03, 2010)
 ==================